refactor: split tests into success and failed snap files.
